feat(appeal): review suggestions

+++ internal/store/postgres/migrations/000020_add_appeal_approval_revision.up.sql
@@ -3,12 +3,12 @@ BEGIN;
 ALTER TABLE
     "appeals"
 ADD
-    COLUMN IF NOT EXISTS "revision" bigint;
+    COLUMN IF NOT EXISTS "revision" int;
 
 ALTER TABLE
     "approvals"
 ADD 
-    COLUMN IF NOT EXISTS "appeal_revision" bigint,
+    COLUMN IF NOT EXISTS "appeal_revision" int,
 ADD 
     COLUMN IF NOT EXISTS "is_stale" boolean;
